requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
requests==2.31.0
gpxpy>=1.6.0
numpy>=1.24.0
matplotlib>=3.7.0
shapely>=2.1.0
polyline>=1.5.0
python-dotenv>=1.0.0
python-multipart>=0.0.6